requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11
torch==2.6.0
langchain-core==0.3.58
langchain-community==0.3.23
langchain-openai==0.3.16
langchain-qdrant==0.2.0
qdrant-client==1.14.2
FlagEmbedding==1.3.4
langsmith==0.3.42
python-dotenv==1.1.0